Add feign
Feign is a java to http client binder inspired by Retrofit, JAXRS-2.0, and WebSocket.
Feign's first goal was reducing the complexity of binding Denominator uniformly to http apis regardless of restfulness.
I've used spring-cloud netflix which adds spring integration to feign so
you just reuse the normal spring annotations.

Our logging integration has also been configured

It seems really nice so far, and it will fix our issue with not getting
pdf service logs through our logging
