+++ build.gradle
@@ -116,6 +116,9 @@ dependencies {
 
     compile group: 'org.springframework', name: 'spring-context-support', version: '4.3.10.RELEASE'
     compile group: 'org.springframework.retry', name: 'spring-retry', version: '1.2.1.RELEASE'
+    compile group: 'org.springframework.cloud', name: 'spring-cloud-starter-feign', version: '1.3.5.RELEASE'
+
+    compile group: 'io.github.openfeign', name: 'feign-httpclient', version: '9.5.0'
 
     testCompile group: 'org.springframework.boot', name: 'spring-boot-starter-test', version: versions.springBoot
     testCompile group: 'org.assertj', name: 'assertj-core', version: '3.8.0'
