+++ src/main/java/uk/gov/hmcts/cmc/claimstore/ClaimStoreApplication.java
@@ -3,9 +3,11 @@ package uk.gov.hmcts.cmc.claimstore;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
+import org.springframework.cloud.netflix.feign.EnableFeignClients;
 
 @SpringBootApplication(exclude = DataSourceAutoConfiguration.class)
 @SuppressWarnings({"HideUtilityClassConstructor", "squid:S1118"}) // Spring needs a constructor, its not a utility class
+@EnableFeignClients
 public class ClaimStoreApplication {
 
     public static final String BASE_PACKAGE_NAME = ClaimStoreApplication.class.getPackage().getName();
