+++ src/main/java/uk/gov/hmcts/cmc/claimstore/config/FeignClientConfiguration.java
@@ -1,0 +1,35 @@
+package uk.gov.hmcts.cmc.claimstore.config;
+
+import feign.Client;
+import feign.httpclient.ApacheHttpClient;
+import org.apache.http.HttpRequestInterceptor;
+import org.apache.http.HttpResponseInterceptor;
+import org.apache.http.client.config.RequestConfig;
+import org.apache.http.impl.client.HttpClientBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import uk.gov.hmcts.reform.logging.httpcomponents.OutboundRequestIdSettingInterceptor;
+import uk.gov.hmcts.reform.logging.httpcomponents.OutboundRequestLoggingInterceptor;
+
+@Configuration
+public class FeignClientConfiguration {
+
+    @Bean
+    public Client getFeignHttpClient() {
+        int timeout = 10000;
+        RequestConfig config = RequestConfig.custom()
+            .setConnectTimeout(timeout)
+            .setConnectionRequestTimeout(timeout)
+            .setSocketTimeout(timeout)
+            .build();
+
+        return new ApacheHttpClient(HttpClientBuilder
+            .create()
+            .useSystemProperties()
+            .addInterceptorFirst(new OutboundRequestIdSettingInterceptor())
+            .addInterceptorFirst((HttpRequestInterceptor) new OutboundRequestLoggingInterceptor())
+            .addInterceptorLast((HttpResponseInterceptor) new OutboundRequestLoggingInterceptor())
+            .setDefaultRequestConfig(config)
+            .build());
+    }
+}
