+++ src/main/java/uk/gov/hmcts/cmc/claimstore/idam/IdamApi.java
@@ -1,0 +1,22 @@
+package uk.gov.hmcts.cmc.claimstore.idam;
+
+import org.springframework.cloud.netflix.feign.FeignClient;
+import org.springframework.http.HttpHeaders;
+import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import uk.gov.hmcts.cmc.claimstore.idam.models.GeneratePinRequest;
+import uk.gov.hmcts.cmc.claimstore.idam.models.GeneratePinResponse;
+import uk.gov.hmcts.cmc.claimstore.idam.models.UserDetails;
+
+@FeignClient(name = "idam-api", url = "${idam.api.url}")
+public interface IdamApi {
+    @RequestMapping(method = RequestMethod.GET, value = "/details")
+    UserDetails retrieveUserDetails(@RequestHeader(HttpHeaders.AUTHORIZATION) String authorisation);
+
+    @RequestMapping(method = RequestMethod.POST, value = "/pin")
+    GeneratePinResponse generatePin(
+        GeneratePinRequest requestBody,
+        @RequestHeader(HttpHeaders.AUTHORIZATION) String authorisation
+    );
+}
