+++ src/main/java/uk/gov/hmcts/cmc/claimstore/services/WorkingDayIndicator.java
@@ -1,6 +1,7 @@
 package uk.gov.hmcts.cmc.claimstore.services;
 
 import org.springframework.stereotype.Component;
+import uk.gov.hmcts.cmc.claimstore.services.bankholidays.PublicHolidaysCollection;
 
 import java.time.DayOfWeek;
 import java.time.LocalDate;
