+++ src/test/java/uk/gov/hmcts/cmc/claimstore/services/WorkingDayIndicatorTest.java
@@ -5,6 +5,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
+import uk.gov.hmcts.cmc.claimstore.services.bankholidays.PublicHolidaysCollection;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
